AIagent/MyColabProject/social_media_generator/backend/agents/analyzer_agent.py
# backend/agents/analyzer_agent.py
from autogen import AssistantAgent

class SocialMediaAnalyzerAgent(AssistantAgent):
    """Analyzes and predicts post performance"""

    def __init__(self, **kwargs):
        system_message = """You are a Social Media Analytics Expert who can predict post performance and provide optimization recommendations.

Analyze the complete post package (strategy, copy, visual direction, platform optimizations) and provide:

1. ENGAGEMENT_PREDICTION: Likely performance on each platform (high/medium/low) with reasoning
2. OPTIMIZATION_SUGGESTIONS: Specific improvements to increase engagement
3. TIMING_RECOMMENDATIONS: Best posting times for target audience
4. HASHTAG_ANALYSIS: Assessment of hashtag strategy effectiveness
5. VIRAL_POTENTIAL: Elements that could make this content go viral
6. RISK_ASSESSMENT: Any potential issues or concerns
7. A/B_TEST_IDEAS: Specific variations to test for better performance

Provide actionable, data-driven insights based on current social media trends and best practices.

Format your response with clear sections for each analysis area.
"""
