AIagent/MyColabProject/social_media_generator/requirements.txt
streamlit>=1.28.0
autogen-agentchat[lmm]~=0.2
diskcache>=5.6.1
openai>=1.3.0
matplotlib>=3.7.0
pillow>=10.0.0
termcolor>=2.3.0
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.0.0
requests>=2.31.0
